+++ NEWS.md
@@ -1,5 +1,6 @@
 ## shoredate 1.0.1.9000 (development version)
- - `shoredate_hdr()` now returns the weighted arithmetic mean.
+ - `shoredate_hdr()` and `shoreline_date()` now returns the weighted mean date.
+ - `shoreline_date()` now has the parameter model, which accepts either "none" or "gamma". If "none" is specified then the distance between site and shoreline is assumed to be zero.
 
 ## shoredate 1.0.1 (2023-03-06)